Mixin component. Use Sale for Payment Action for correct create invoice.
